requirements.txt
matplotlib==3.1.1
tensorflow==2.2.0
numpy==1.17.2
h5py==2.10.0
dlib==19.20.0
opencv_python==4.2.0.32
imutils==0.5.3
pandas==0.25.1
Keras==2.4.3
tqdm==4.61.0
scipy==1.4.1
Pillow==8.2.0
